requirements.txt
pytest
selenium
pytest-xdist
pytest-bdd
webdriver_manager
allure-pytest